2490: update solution

+++ algorithm_2000-2999/2490-Circular Sentence/main.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isCircularSentence(string sentence) {
-        for (size_t i = 0; i < sentence.size(); ++i) {
+        for (int i = 0; i < sentence.size(); ++i) {
             if (sentence[i] == ' ' && sentence[i - 1] != sentence[i + 1]) {
                 return false;
             }
